Loads of changes
Added Toolbox for quick visual toggles.
TODO: Make buttons reflect toggle state

Fixed AppOverlay not really working all that well.

Added some descriptions and tool tips

Various under the hood changes.
A ton of the functions in the main form are now Public. This was done in order to be lazy and not rewrite things for the toolbox. Ideally this should also allow people to theoretically hook into them a bit easier for customizations.