Inversion Testing
Testing more inversion in the block/tile filters.
Added the Windows User Action Hook library to the program to capture mouse scrolling. Is this overkill? yeah. I'll try to downsize the added library into maybe just a simple class or something so we dont have to haul around a DLL.

Scrolling should temporarily disable the inversion and reenable it after 500 miliseconds or no scroll detection.